feat: added alice_test, updated alice_dio tests
