+++ packages/alice_dio/pubspec.yaml
@@ -25,3 +25,4 @@ dev_dependencies:
   test: ^1.25.2
   mocktail: ^1.0.4
   http_mock_adapter: ^0.6.1
+  alice_test: ^1.0.0
