+++ packages/alice_test/README.md
@@ -1,0 +1,3 @@
+# Alice Test
+
+Package which exposes test utils for all Alice packages.
