+++ packages/alice_test/analysis_options.yaml
@@ -1,0 +1,10 @@
+analyzer:
+  exclude:
+    - "lib/generated_plugin_registrant.dart"
+    - "**.g.dart"
+
+include: package:flutter_lints/flutter.yaml
+
+linter:
+  rules:
+  #library_annotations: false
