+++ packages/alice_test/lib/alice_test.dart
@@ -1,0 +1,4 @@
+export 'matcher/header_matcher.dart';
+export 'matcher/alice_call_matcher.dart';
+export 'matcher/alice_request_matcher.dart';
+export 'matcher/alice_response_matcher.dart';
