+++ packages/alice_test/pubspec.yaml
@@ -1,0 +1,18 @@
+name: alice_test
+description: "Test utils for Alice packages."
+version: 1.0.0
+repository: https://github.com/jhomlala/alice
+homepage: https://github.com/jhomlala/alice
+publish_to: none
+
+environment:
+  sdk: ^3.0.0
+
+dependencies:
+  alice: ^1.0.0-dev.9
+  test: ^1.25.2
+
+dev_dependencies:
+  flutter_test:
+    sdk: flutter
+  flutter_lints: ^4.0.0
